backend/requirements.txt
# Core Data
pandas
numpy

# Database
pyodbc

# Google & AI
google-generativeai
google-api-python-client

# Profanity Filter
better-profanity

# Web
requests

# Config
python-dotenv
azure-storage-blob

# Machine Learning
tensorflow==2.16.1
scikit-learn==1.5.0